Praktikum 4,5, dan 6

+++ H071191054/Praktikum_3/Faktorial.java
@@ -5,7 +5,7 @@ class Faktorial {
         Scanner sl = new Scanner(System.in);
         System.out.print("Input jumlah deret faktorial = ");
         int x = sl.nextInt();
-        long result = 1;
+        int result = 1;
         System.out.printf("%d! = ", x);
         while (x > 1) {
             System.out.print(x + " * ");
